Trying to find the Forward Delete keycode
There is some misalignment with the keycodes. In the main ZMK firmware
docs, they use BACKSPACE and DEL. BACKSPACE is also known as DELETE and
DEL is known as FORWARD DELETE. However, there is some kind of bug in
the firmware where BACKSPACE does nothing and DEL is BACKSPACE.

So now I'm trying to find out what is actually FORWARD DELETE becuase I
do use it (esp on windows).